Create CutRod.java
Given a rod of length n inches and an array of prices that contains prices of all pieces of size smaller than n. Determine the maximum value obtainable by cutting up the rod and selling the pieces. For example - 
lengths = {1,2,3,4,5,6,7,8,9,10}
price = {10, 50, 80, 90, 100, 170, 170, 200, 240, 300}
Find the optimal solution to cut the rod to maximize the revenue.